+++ src/CommandBus/Attribute/AsCommandHandler.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * SPDX-FileCopyrightText: 2020 Communitales GmbH
+ *
+ * SPDX-License-Identifier: MIT
+ */
+
+namespace Communitales\Component\CommandBus\Attribute;
+
+use Attribute;
+
+#[Attribute(Attribute::TARGET_CLASS)]
+final readonly class AsCommandHandler
+{
+    /**
+     * @param class-string $handles
+     */
+    public function __construct(public string $handles)
+    {
+    }
+}