+++ src/CommandBus/Handler/CommandHandlerInterface.php
@@ -1,22 +1,21 @@
 <?php
-/**
- * @copyright   Copyright (c) 2020 - 2023 Communitales GmbH (https://www.communitales.com/)
+
+declare(strict_types=1);
+
+/*
+ * SPDX-FileCopyrightText: 2020 Communitales GmbH
  *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
+ * SPDX-License-Identifier: MIT
  */
 
 namespace Communitales\Component\CommandBus\Handler;
 
 use Communitales\Component\CommandBus\Command\CommandInterface;
-use Communitales\Component\CommandBus\Handler\Result\CommandHandlerResultInterface;
+use Communitales\Component\CommandBus\Handler\Result\CommandResultInterface;
 
-/**
- * Interface CommandHandlerInterface
- */
+/** @template-contravariant TCommand of CommandInterface */
 interface CommandHandlerInterface
 {
-    public function canHandle(CommandInterface $command): bool;
-
-    public function handle(CommandInterface $command): CommandHandlerResultInterface;
+    /** @param TCommand $command */
+    public function handle(CommandInterface $command): CommandResultInterface;
 }