+++ src/CommandBus/Handler/Result/CommandResultException.php
@@ -1,9 +1,11 @@
 <?php
-/**
- * @copyright   Copyright (c) 2020 - 2024 Communitales GmbH (https://www.communitales.com/)
+
+declare(strict_types=1);
+
+/*
+ * SPDX-FileCopyrightText: 2020 Communitales GmbH
  *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
+ * SPDX-License-Identifier: MIT
  */
 
 namespace Communitales\Component\CommandBus\Handler\Result;
@@ -12,16 +14,16 @@ use RuntimeException;
 use Throwable;
 
 /**
- * An exception representing an CommandHandlerResult.
- * This is useful if you want to return a result from a sub method of a handler method.
+ * An exception carrying a command result.
+ * This is useful for returning a result early from a nested handler method.
  */
 class CommandResultException extends RuntimeException
 {
     public function __construct(
-        public readonly CommandHandlerResultInterface $commandResult,
+        public readonly CommandResultInterface $commandResult,
         string $message = '',
         int $code = 0,
-        Throwable $previous = null
+        ?Throwable $previous = null
     ) {
         parent::__construct($message, $code, $previous);
     }