+++ tests/Unit/Handler/TestCommand.php
@@ -1,10 +1,11 @@
 <?php
 
-/**
- * @copyright   Copyright (c) 2024 Communitales GmbH (https://www.communitales.com/)
+declare(strict_types=1);
+
+/*
+ * SPDX-FileCopyrightText: 2020 Communitales GmbH
  *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
+ * SPDX-License-Identifier: MIT
  */
 
 namespace Communitales\Test\Unit\Component\CommandBus\Handler;