+++ tests/Unit/Handler/TestCommandHandler.php
@@ -1,52 +1,40 @@
 <?php
 
-/**
- * @copyright   Copyright (c) 2024 Communitales GmbH (https://www.communitales.com/)
+declare(strict_types=1);
+
+/*
+ * SPDX-FileCopyrightText: 2020 Communitales GmbH
  *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
+ * SPDX-License-Identifier: MIT
  */
 
 namespace Communitales\Test\Unit\Component\CommandBus\Handler;
 
+use Communitales\Component\CommandBus\Attribute\AsCommandHandler;
 use Communitales\Component\CommandBus\Command\CommandInterface;
 use Communitales\Component\CommandBus\Handler\CommandHandlerInterface;
-use Communitales\Component\CommandBus\Handler\CommandHandlerTrait;
-use Communitales\Component\CommandBus\Handler\Result\CommandHandlerResultInterface;
-use Communitales\Component\CommandBus\Handler\Result\SuccessResult;
+use Communitales\Component\CommandBus\Handler\Result\CommandResult;
+use Communitales\Component\CommandBus\Handler\Result\CommandResultInterface;
 use Communitales\Component\StatusBus\StatusMessage;
-use Override;
-use Symfony\Component\Translation\TranslatableMessage;
 
 /**
  * Class TestCommandHandler
  */
+/** @implements CommandHandlerInterface<TestCommand> */
+#[AsCommandHandler(TestCommand::class)]
 class TestCommandHandler implements CommandHandlerInterface
 {
-    use CommandHandlerTrait;
-
-    #[Override]
-    public function canHandle(CommandInterface $command): bool
-    {
-        return $command instanceof TestCommand;
-    }
-
-    public function test(TestCommand $command): CommandHandlerResultInterface
+    /** @param TestCommand $command */
+    public function handle(CommandInterface $command): CommandResultInterface
     {
         if ($command->test === 'success') {
-            return new SuccessResult(
-                StatusMessage::createSuccessMessage(
-                    new TranslatableMessage(
-                        $command->test
-                    )
-                )
+            return CommandResult::success(
+                StatusMessage::success($command->test)
             );
         }
 
-        return new SuccessResult(
-            StatusMessage::createErrorMessage(
-                $command->test
-            )
+        return CommandResult::error(
+            StatusMessage::error($command->test)
         );
     }
 }